# groupomania-openapi-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "groupomania-api",
    "description": "",
    "license": {
      "name": "ISC"
    }
  },
...
  "servers": [
    {
      "url": "https://localhost:3000"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Finds a post by ID",
        "description": "Returns the post with the given id",
        "parameters": [
          {
            "$ref": "#/components/parameters/postIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "posts"
        ],
        "summary": "Modification of a post",
        "description": "Update the post with the id submitted. If an image is uploaded, it is captured, and the imageUrl of the post is updated. If no file is submitted, the informations of the post are in the root of the request body. If a file is submitted, the post (string) is in req.body.post The initial request body is empty ; multer returns a string for the request body with the data submitted with the file.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/postIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/post-form"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "post": {
                    "$ref": "#/components/schemas/post-form-with-image"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "image to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/post"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "Deletion of a post",
        "description": "Delete the post with the given id",
        "parameters": [
          {
            "$ref": "#/components/parameters/postIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Post deletion message"
                },
                "example": "Post deleted"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "postIdParam": {
        "description": "post ID",
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "description": "ID of the post"
        }
      }
    },
    "schemas": {
      "errorMessage": {
        "type": "string",
        "maxLength": 256,
        "description": "A text message describing an error",
        "example": "Error during the request processing"
      },
...
      "post": {
        "type": "object",
        "description": "data relative to a post (object structure -> post schema)",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the post"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "description": "Date of creation of the post"
          },
          "updatedAt": {
            "type": "string",
            "format": "date",
            "description": "Date of update of the post"
          },
          "text": {
            "type": "string",
            "description": "Content (text) of the post"
          },
          "imageUrl": {
            "type": "string",
            "description": "Content (image URL) of the post"
          },
          "imageAlt": {
            "type": "string",
            "description": "Content (image URL) of the post"
          },
          "authorId": {
            "type": "string",
            "description": "ID of the user that created the post"
          },
          "author": {
            "type": "object",
            "description": "User that created the post",
            "properties": {
              "firstName": {
                "type": "string",
                "description": "First name of the user that created the post"
              },
              "surName": {
                "type": "string",
                "description": "Surname of the user that created the post"
              },
              "pseudo": {
                "type": "string",
                "description": "Pseudonym of the user that created the post"
              }
            },
            "required": [
              "firstName",
              "surName"
            ]
          },
          "likes": {
            "type": "array",
            "description": "Array of users (user IDs) that likes the post",
            "items": {
              "type": "object",
              "description": "User that created the post",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "ID of a user that likes the post"
                }
              }
            },
            "minItems": 0
          },
          "dislikes": {
            "type": "array",
            "description": "Array of users (user IDs) that dislikes the post",
            "items": {
              "type": "object",
              "description": "User that created the post",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "ID of a user that dislikes the post"
                }
              }
            },
            "minItems": 0
          },
          "comments": {
            "type": "array",
            "description": "Array of comments of the post",
            "items": {
              "$ref": "#/components/schemas/comment"
            },
            "minItems": 0
          }
        },
        "example": {
          "id": "cl6dl8yq10022b4muffz9isxl",
          "createdAt": "2022-08-03T12:31:30.505Z",
          "updatedAt": "2022-08-03T12:31:30.505Z",
          "text": "This is a post text",
          "imageUrl": "http://localhost:3000/images/Post1_image_20220701_162713.jpg",
          "imageAlt": "This is the description of the image",
          "authorId": "e5268c386c9b17c39bd6a17d",
          "likes": [
            {
              "id": "c386c9b17c39be5268d6a17d"
            },
            {
              "id": "9b17c39be5268dc386c6a17d"
            }
          ],
          "dislikes": [
            {
              "id": "be5268dc386c6a9b17c3917d"
            },
            {
              "id": "a9b17c3917dbe5268dc386c6"
            }
          ],
          "comments": [
            {
              "id": "cl977p9mb000at2vkmudnbxcp",
              "createdAt": "2022-08-03T12:31:30.505Z",
              "text": "This is a comment",
              "authorId": "e5268c386c9b17c39bd6a17d",
              "postId": "cl6dl8yq10022b4muffz9isxl"
            }
          ]
        }
      },
      "post-form": {
        "type": "object",
        "description": "data to use to create the post without image (object structure -> post-form schema)",
        "properties": {
          "text": {
            "type": "string",
            "description": "Content (text) of the post"
          },
          "authorId": {
            "type": "string",
            "description": "ID of the user that created the post"
          }
        },
        "example": {
          "text": "This is a post text",
          "authorId": "e5268c386c9b17c39bd6a17d"
        }
      },
      "post-form-with-image": {
        "type": "object",
        "description": "data to use to create the post with image (object structure -> post-form-with-image schema)",
        "properties": {
          "text": {
            "type": "string",
            "description": "Content (text) of the post"
          },
          "authorId": {
            "type": "string",
            "description": "ID of the user that created the post"
          },
          "imageAlt": {
            "type": "string",
            "description": "Description (alternative text) of the image (required if an image is provided)"
          }
        },
        "example": {
          "text": "This is a post text",
          "authorId": "e5268c386c9b17c39bd6a17d",
          "imageAlt": "This is the description of the image"
        }
      },
      "comment": {
        "type": "object",
        "description": "data relative to a comment (object structure -> comment schema)",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the comment"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation date of the comment"
          },
          "text": {
            "type": "string",
            "description": "Text of the comment"
          },
          "authorId": {
            "type": "string",
            "description": "ID of the author of the comment"
          },
          "postId": {
            "type": "string",
            "description": "ID of the post of the comment"
          }
        },
        "example": {
          "id": "cl977p9mb000at2vkmudnbxcp",
          "createdAt": "2022-08-03T12:31:30.505Z",
          "text": "This is a comment",
          "authorId": "e5268c386c9b17c39bd6a17d",
          "postId": "cl6dl8yq10022b4muffz9isxl"
        }
      }
    }
  }
}